Now "forking" a generic version of the SnP500 one
The same as https://github.com/EliteLoser/Finance/blob/master/SnP500/Get-SnP500PeriodicDevelopment.ps1 as of 2019-02-24, but now it's generic, and the function name is Get-PeriodicDevelopment.

Keeping SnP500 examples. Dynamic titles for other data sources than '^GSPC.csv'.

It can be used to predict when a disk will reach zero for that matter... Should present/"market" it in different useful ways... I've written code to extract SCOM data warehouse data from the 30 and 180 day stores and make predictions. We caught quite a bit so far. Remains to automate the reporting, or rather what happens AFTER the reporting to Splunk (formalized processes are harder than code...).